Modify getOpts()
I removed the -m option, because it needed a separate menu to be convenient,
which went against the KISS philosophy of this app.
Use -d to simply download music with yt-dlp and use it,
Why make it complicated when you can organize your music in directories? :)
musicPath is again a (char**).